fix(ssm,quadrature): preserve input dtype under x64 (gh-219) (#221)
* fix(ssm,quadrature): preserve input dtype under x64 (gh-219)

Two independent float64 promotions made float32 unusable end to end when
jax_enable_x64 is on. Pure float32 (x64 off) was unaffected, so this bit
mixed-precision users rather than the default path.

kalman_filter's predict-only branch returned a bare jnp.array(0.0) — a
float64 scalar under x64 — while the update branch's accumulator followed
the float32 inputs, so lax.cond rejected the branches outright. Carry the
dtype through in _skip_update and init_carry, as nonlinear_kalman_filter
already does.

The quadrature rules built their points and weights from Python and NumPy
scalars, landing in float64 regardless of the input dtype and promoting
every downstream moment. Build them in mean.dtype throughout:

- sigma_points, cubature_points, fifth_order_cubature_points: weights
- fifth_order_cubature_points: also the shell points, whose np.sqrt radii
  were strongly-typed np.float64 and promoted what they scaled
- gauss_hermite_points: takes an explicit dtype, since it has no mean to
  key off; GaussHermiteIntegrator passes the input dtype
- MonteCarloIntegrator, AssumedDensityFilter: draws, weights, and jitter

uv.lock picks up the 0.0.19 -> 0.0.21 version sync it had drifted from.


* chore: drop incidental uv.lock version sync from this PR

The 0.0.19 -> 0.0.21 lockfile sync is release metadata, unrelated to the
dtype fix. Per AGENTS.md's surgical-change guidance it belongs in its own
change, not bundled here.


---------

+++ src/gaussx/_quadrature/_adf.py
@@ -84,7 +84,7 @@ class AssumedDensityFilter(AbstractIntegrator):
 
         # Sample from input Gaussian
         L = cholesky(state.cov).as_matrix()
-        eps = jr.normal(key, (self.n_samples, N))
+        eps = jr.normal(key, (self.n_samples, N), dtype=mu.dtype)
         x_samples = mu[None, :] + eps @ L.T
 
         # Propagate samples
@@ -101,7 +101,7 @@ class AssumedDensityFilter(AbstractIntegrator):
             eps_reg = self.regularization * jnp.trace(Sigma_y) / M
         else:
             eps_reg = self.regularization
-        Sigma_y = Sigma_y + eps_reg * jnp.eye(M)
+        Sigma_y = Sigma_y + eps_reg * jnp.eye(M, dtype=Sigma_y.dtype)
         Sigma_y = symmetrize(Sigma_y)
 
         # Cross-covariance
